# Regress daemon