--- a/aha/util/pnr.py
+++ b/aha/util/pnr.py
@@ -119,14 +119,22 @@ def dispatch(args, extra_args=None):
     # When running as daemon, must use non-blocking "Popen" and not "check_call"
     # if '--daemon' in extra_args and not 'use' in extra_args:
 
+    print(f"extra_args={extra_args}", flush=True)
+
     launch_daemon = False
     do_cmd = subprocess.check_call
     if '--daemon' in extra_args:
+        print(f"--- found the --daemon", flush=True)
+        print(f"--- does daemon exist yet?", flush=True)
         cmd = [sys.executable, "garnet.py", "--daemon", "status"]
         p = subprocess.run(cmd, text=True,
             stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
         launch_daemon = 'no daemon found' in p.stdout
-        if launch_daemon: do_cmd = subprocess.Popen
+        if launch_daemon:
+            print(f"--- found no daemon, setting do_cmd to Popen", flush=True)
+            do_cmd = subprocess.Popen
+        else:
+            print(f"--- found the daemon, leaving do_cmd alone", flush=True)
 
     subprocess_call_log (
         cmd=[sys.executable, "garnet.py"] + map_args + extra_args,