--- a/aha/util/regress.py
+++ b/aha/util/regress.py
@@ -148,6 +148,7 @@ def test_dense_app(test, width, height, layer=None, env_parameters=""):
             test,
             "--width", str(width),
             "--height", str(height),
+            "--daemon auto",
             "--env-parameters", env_parameters,
         ] + layer_array
     )
@@ -162,6 +163,7 @@ def test_dense_app(test, width, height, layer=None, env_parameters=""):
 
 
 def dispatch(args, extra_args=None):
+    print(f"--- extra_args={extra_args}", flush=True)
     sparse_tests = []
     if args.config == "fast":
         width, height = 4, 4