This pull enables an option to use the new daemon-enabled `garnet.py` to speed regression tests. This option is currently ON by default. We can open a discussion if one or more reviewer(s) thinks that should be different...

Code changes include:
- updated `pnr.py` to use background daemon instead of doing pnr from scratch every time;
- updated `regress.py` to use this new `pnr.py`;
- added frequent `stdout flush` commands to try and improve stdout/stderr synchronization in regression log output.

The daemon saves approximately 90 seconds per dense app in a regression suite. (Sparse apps use a different mechanism for pnr.) Thus, for a benchmark suite of multiple apps, the time savings is not large -- the daemon was designed mainly to help trunaround time for interactive testing. But, including the daemon in the test suite gives us a mechanism to test and maintain it going forward.

Below are actual comparisons for 4-hour "daily" and 2-hour "pr" regression suites, with and without using the daemon.

### "Daily" regression suite
Comparing times (in seconds) for some of the dense apps in the "daily" suite.

BEFORE/daily: `onyx integration tests` step takes 3h 47m
https://buildkite.com/stanford-aha/aha-flow/builds/9602

```
step                         total    compile       map       test
---------------------------- -----  ---------  --------  ---------
apps/gaussian_glb              427        165        207        53
apps/pointwise_glb             301         67        197        36
apps/unsharp_glb               584        160        222       200
apps/camera_pipeline_2x2_glb  1200        337        368       494
apps/harris_color_glb          487        179        222        85
apps/cascade_glb               304         68        195        41
apps/maxpooling_glb           2616        305       1122      1188
```

AFTER/daily: `onyx integration tests` step takes 3h 40m
https://buildkite.com/stanford-aha/aha-flow/builds/9632
```
apps/gaussian_glb              477        174        249        53
apps/pointwise_glb             200         69         91        39
apps/unsharp_glb               523        164        126       232
apps/camera_pipeline_2x2_glb  1147        350        272       524
apps/harris_color_glb          409        188        131        89
apps/cascade_glb               222         73        105        42
apps/maxpooling_glb           2706        318       1068      1319
```

Comparing "map" times, we see that daemon saves about 90 seconds from each dense app:
```
                                D     No D     diff
apps/gaussian_glb             207      249      +41
apps/pointwise_glb            197       91     -105
apps/unsharp_glb              222      126      -96
apps/camera_pipeline_2x2_glb  368      272      -96
apps/harris_color_glb         222      131      -91
apps/cascade_glb              195      105      -89
apps/maxpooling_glb          1122     1068      -53
```
### "PR" regression suite
Similarly, we can compare times (in seconds) for the 'pr' regression suite.

BEFORE/pr: `onyx integration tests` step takes 1h 40m
https://buildkite.com/stanford-aha/aha-flow/builds/9604

AFTER/pr `onyx integration tests` step takes 1h 25m
https://buildkite.com/stanford-aha/aha-flow/builds/9676

The pr suite contains eleven dense apps, so we expect about 15 minutes' saving, which is what we got.