@steveri 2024-01-10
With Kalhan's help, I merged this code with the recently-merged aha fix. In addition, we added code to kill any existing daemons when/if `aha regress` builds new verilog. Once the tests have passed, I will finalize the pull, unless someone wants extra time to look over the new changes...