@kalhankoul96 2024-01-10
Just curious, what was the issue before when I was in gates?