@steveri 2024-01-11
> Just curious, what was the issue before when I was in gates?

Haha, turns out that "fast" suite has two apps, one sparse and one dense. Because VCS was not loaded, the sparse app failed, so the dense app never launched. Only dense apps use the daemon, so...no daemon. Once I got the VCS working and passing, the sparse app passed and then the dense app launched the daemon without further drama...